refactor disruption checks to separate module
