+++ tests/test_main.py
@@ -2,95 +2,23 @@ import sys
 import os
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 import pytest
-from main import (
-    parse_cache_control,
-    disruption_hash,
-    extract_all_disruptions,
-    disruption_id_hash,
-)
+from unittest.mock import patch
 
-@pytest.mark.parametrize(
-    "headers,expected",
-    [
-        ({"Cache-Control": "public, must-revalidate, max-age=30, s-maxage=60"}, 30),
-        ({"Cache-Control": "max-age=120"}, 120),
-        ({"Cache-Control": "no-cache"}, 60),
-        ({}, 60),
-        ({"Cache-Control": "max-age=notanint"}, 60),
-        ({"Cache-Control": "public, max-age=45, foo=bar"}, 45),
-        ({"Cache-Control": "public, s-maxage=60"}, 60),
-        ({"Cache-Control": "max-age=0"}, 0),
-    ],
-)
-def test_parse_cache_control(headers, expected):
-    assert parse_cache_control(headers) == expected
+from main import notify_on_disruption_change
 
-def test_disruption_hash_changes():
-    d1 = [{"lineId": "central", "reason": "Minor delays"}]
-    d2 = [{"lineId": "central", "reason": "Good Service"}]
-    d3 = [{"reason": "Minor delays", "lineId": "central"}]
-    # Different disruptions should have different hashes
-    assert disruption_hash(d1) != disruption_hash(d2)
-    # Identical disruptions (even as new dicts) should have the same hash
-    assert disruption_hash(d1) == disruption_hash([{**d1[0]}])
-    # Hash should be consistent regardless of order
-    assert disruption_hash(d1) == disruption_hash(d3)
-
-def test_disruption_hash_empty_list():
-    # disruption_hash should handle empty list
-    assert isinstance(disruption_hash([]), str)
-    # Hash of empty list should be consistent
-    assert disruption_hash([]) == disruption_hash([])
-
-def test_disruption_hash_multiple_disruptions():
-    d1 = [
+def test_notify_on_disruption_change():
+    disruptions = [
+        {"lineId": "piccadilly", "reason": "Closed for maintenance"},
         {"lineId": "central", "reason": "Minor delays"},
-        {"lineId": "district", "reason": "Major delays"},
     ]
-    d2 = [
-        {"lineId": "central", "reason": "Minor delays"},
-        {"lineId": "district", "reason": "Major delays"},
-    ]
-    d3 = [
-        {"lineId": "central", "reason": "Minor delays"},
-        {"lineId": "district", "reason": "Good Service"},
-    ]
-    # Hash should be the same for identical lists
-    assert disruption_hash(d1) == disruption_hash(d2)
-    # Hash should differ if any disruption changes
-    assert disruption_hash(d1) != disruption_hash(d3)
-
-def test_extract_all_disruptions():
-    status_json = [
-        {"id": "central", "lineStatuses": [
-            {"statusSeverity": 10, "reason": "All good"},
-            {"statusSeverity": 9, "reason": "Minor delays", "lineId": "central"},
-            {"statusSeverity": 9, "reason": "Minor delays", "lineId": "central"},
-            {"statusSeverity": 8, "reason": "Major delays", "lineId": "central"},
-        ]},
-        {"id": "piccadilly", "lineStatuses": [
-            {"statusSeverity": 7, "reason": "Closed", "lineId": "piccadilly"},
-        ]},
-    ]
-    disruptions = extract_all_disruptions(status_json)
-    expected = [
-        {"statusSeverity": 9, "reason": "Minor delays", "lineId": "central"},
-        {"statusSeverity": 8, "reason": "Major delays", "lineId": "central"},
-        {"statusSeverity": 7, "reason": "Closed", "lineId": "piccadilly"},
-    ]
-    assert {(d["lineId"], d["reason"]) for d in disruptions} == {(d["lineId"], d["reason"]) for d in expected}
-
-@pytest.mark.parametrize(
-    "disruption1,disruption2,should_match",
-    [
-        ( {"lineId": "central", "reason": "Minor delays"}, {"lineId": "central", "reason": "Minor delays"}, True ),
-        ( {"lineId": "central", "reason": "Minor delays"}, {"lineId": "central", "reason": "Major delays"}, False ),
-        ( {"lineId": "central", "reason": "Minor delays"}, {"lineId": "piccadilly", "reason": "Minor delays"}, False ),
-        ( {"lineId": "central", "reason": "Minor delays"}, {"lineId": "central"}, False ),
-    ]
-)
-def test_disruption_id_hash(disruption1, disruption2, should_match):
-    h1 = disruption_id_hash(disruption1)
-    h2 = disruption_id_hash(disruption2)
-    assert (h1 == h2) == should_match
+    
+    # Mock the notifier to capture output
+    mock_notify = patch('notify.notifier.notify_users').start()
+    mock_notify.return_value = None  # Mock the return value
+    # Call the function with mocked notifier
+    notify_on_disruption_change(disruptions, mock_notify)
+    # Assert that the notifier was called with the disruptions
+    assert mock_notify.called
+    assert mock_notify.call_count == 1
+    assert mock_notify.call_args[0][0] == disruptions
